+++ netbox_ssl/api/serializers/__init__.py
@@ -1,4 +1,5 @@
 from .assignments import CertificateAssignmentSerializer
+from .certificate_authorities import CertificateAuthoritySerializer
 from .certificates import CertificateImportSerializer, CertificateSerializer
 from .csr import CertificateSigningRequestSerializer, CSRImportSerializer
 
@@ -6,6 +7,7 @@ __all__ = [
     "CertificateSerializer",
     "CertificateImportSerializer",
     "CertificateAssignmentSerializer",
+    "CertificateAuthoritySerializer",
     "CertificateSigningRequestSerializer",
     "CSRImportSerializer",
 ]