+++ netbox_ssl/api/urls.py
@@ -9,6 +9,7 @@ from . import views
 router = NetBoxRouter()
 router.register("certificates", views.CertificateViewSet)
 router.register("assignments", views.CertificateAssignmentViewSet)
+router.register("certificate-authorities", views.CertificateAuthorityViewSet)
 router.register("csrs", views.CertificateSigningRequestViewSet)
 
 urlpatterns = router.urls