+++ netbox_ssl/filtersets/__init__.py
@@ -1,9 +1,11 @@
 from .assignments import CertificateAssignmentFilterSet
+from .certificate_authorities import CertificateAuthorityFilterSet
 from .certificates import CertificateFilterSet
 from .csr import CertificateSigningRequestFilterSet
 
 __all__ = [
     "CertificateFilterSet",
     "CertificateAssignmentFilterSet",
+    "CertificateAuthorityFilterSet",
     "CertificateSigningRequestFilterSet",
 ]