+++ netbox_ssl/models/__init__.py
@@ -1,4 +1,9 @@
 from .assignments import CertificateAssignment
+from .certificate_authorities import (
+    DEFAULT_CERTIFICATE_AUTHORITIES,
+    CATypeChoices,
+    CertificateAuthority,
+)
 from .certificates import Certificate, CertificateAlgorithmChoices, CertificateStatusChoices
 from .csr import CertificateSigningRequest, CSRStatusChoices
 
@@ -7,6 +12,9 @@ __all__ = [
     "CertificateStatusChoices",
     "CertificateAlgorithmChoices",
     "CertificateAssignment",
+    "CertificateAuthority",
+    "CATypeChoices",
+    "DEFAULT_CERTIFICATE_AUTHORITIES",
     "CertificateSigningRequest",
     "CSRStatusChoices",
 ]